Auto Creation of Proto v0.5.0

+++ server/server_model_pb.js
@@ -147,7 +147,7 @@ proto.mruv.ServerID.deserializeBinaryFromReader = function(msg, reader) {
     var field = reader.getFieldNumber();
     switch (field) {
     case 1:
-      var value = /** @type {number} */ (reader.readInt32());
+      var value = /** @type {number} */ (reader.readInt64());
       msg.setId(value);
       break;
     default:
@@ -181,7 +181,7 @@ proto.mruv.ServerID.serializeBinaryToWriter = function(message, writer) {
   var f = undefined;
   f = message.getId();
   if (f !== 0) {
-    writer.writeInt32(
+    writer.writeInt64(
       1,
       f
     );
@@ -190,7 +190,7 @@ proto.mruv.ServerID.serializeBinaryToWriter = function(message, writer) {
 
 
 /**
- * optional int32 id = 1;
+ * optional int64 id = 1;
  * @return {number}
  */
 proto.mruv.ServerID.prototype.getId = function() {
